+++ tests/acceptance/install_test.sh
@@ -84,6 +84,26 @@ function test_install_downloads_the_given_version() {
     "$("$installed_bashunit" --env "$TEST_ENV_FILE" --version)"
 }
 
+function test_install_downloads_the_given_version_without_dir() {
+  if [[ "$ACTIVE_INTERNET" -eq 1 ]]; then
+    skip "no internet connection" && return
+  fi
+
+  local installed_bashunit="./lib/bashunit"
+  local output
+  output="$(./install.sh 0.19.0)"
+
+  assert_same \
+    "$(printf "> Downloading a concrete version: '0.19.0'\n> bashunit has been installed in the 'lib' folder")" \
+    "$output"
+
+  assert_file_exists "$installed_bashunit"
+
+  assert_same \
+    "$(printf "\e[1m\e[32mbashunit\e[0m - 0.19.0")" \
+    "$("$installed_bashunit" --env "$TEST_ENV_FILE" --version)"
+}
+
 function test_install_downloads_the_non_stable_beta_version() {
   if [[ "$ACTIVE_INTERNET" -eq 1 ]]; then
     skip "no internet connection" && return